+++ domain/src/main/java/com/devsu/fintech/domain/ports/output/AccountRepositorySPI.java
@@ -2,5 +2,5 @@ package com.devsu.fintech.domain.ports.output;
 
 public interface AccountRepositorySPI {
 
-    boolean hasOpenedAccounts(Integer clientId);
+    boolean hasOpenedAccounts(Long clientId);
 }
