+++ infrastructure/src/main/java/com/devsu/fintech/infrastructure/adapter/jpa/repository/AccountJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountJpaRepository extends JpaRepository<AccountEntity, Long> {
 
-    boolean existsByClientIdAndAccountStatusIdNot(Integer clientId, Integer accountStatusId);
+    boolean existsByClientIdAndAccountStatusIdNot(Long clientId, Integer accountStatusId);
 }
